Refactor enable_touch_scroll_feedback flag
The flag was read-write, but that caused a performance regression (see
b/364725557), so we changed it to read-only. However, the flag ramp
process has been blocked due to this property change.
https://screenshot.googleplex.com/7AqAsAqnHdhtACx
To mitigate this, this change renames the flag, to generate a new flag
ramp process. In addition, we're also changing the namespace to align to
the team that's the primary consumer of the feature, and using a more
specific bug ID.

Flag: android.view.flags.enable_scroll_feedback_for_touch
Test: presubmit
Bug: 382135785
Change-Id: I7adb09dcc6e73996ff69453d770c9fc42f34d5ae