[flexiglass] changeScene even if going to same scene
Removed unnecessary (and incorrect) check from SceneInteactor.validateSceneChange that was causing a bug where the change to Lockscreen triggered by device sleep (power button) while dragging up from Lockscreen to reveal the Bouncer wasn't actually going to the Lockscreen scene.

The check is unnecessary because the upstream logic in STL is already
protecting us and is incorrect because it was using the current scene as
the "from" scene, even though that's not necessarily correct when
transitioning from scene A to scene B.

Fix: 379844479
Test: manually verified that dragging up on the Lockscreen scene, not
removing the finger, and then pressing the power button correctly
transitioned to the Lockscreen scene even though the finger wasn't
removed. Also made sure that additional gestures afterwards work as
intended.
Flag: com.android.systemui.scene_container

Change-Id: I4065506623cac2e33a38004cc4a88dbb56b97b18