Update responsive grid layout + alpha animation
This change updates the layout to be based on the screen size instead of
the available size. This is needed to ensure we always use the same
number of items in landscape/portrait, as the screen size is guaranteed
to not change between orientations. However, the available size can
change depending on the placement of items on the screen.

We also add an alpha animation to hide partial columns and fade them in
as the user scrolls.

Bug: 378171351
Test: manually on folded + unfolded foldable device in both orientations
Flag: com.android.systemui.communal_responsive_grid
Change-Id: I978a881371609dbf194dbfaec6ac23fc0e8b7357