[Notif redesign] Fix broken margins in HUNs
We're updating the top margin in applyStandardTemplateWithActions to
leave space for the top line view in expanded notifications - but it
turns out that this method is also being used for HUNs, which use the
collapsed base template that doesn't need this margin. This lead to
a weird empty space between the title and the content in HUNs, so let's
just not apply this margin for them.

Bug: 378660052
Test: post HUN and check that it looks fine; also make sure normal
collapsed notifs are still okay; screenshot tests to come later
Flag: android.app.notifications_redesign_templates

Change-Id: I8068635c2c88954c70c56cf292e91307052c7890