Use matcher to check keys received in ViewRootImplTest
In this CL, we refactor the ViewRootImplTest to store all of the
received keys in the queue. This will allow richer assertions, like
ensuring that specific keys were received, and in the expected order.

At the same time, move the BlockingQueueEventVerifier into cts-input-lib
so that it can be used by other input-related tests.

Bug: 380522059
Flag: TEST_ONLY
Test: atest ViewRootImplTest
Change-Id: I11dae887ba50db1bfe4726c865d60af2dbf211b2