[flexiglass] Left and right edges of pattern bouncer are touchable
This CL makes drag gestures that start inside the gaps between the
left and right edges of the pattern bouncer pattern area and the edges
of the display still work as legitimate drag gestures.

Before this CL, drag gestures that started in those gaps to the left and
right of the pattern dot matrix did not cause dots to be selected, even
if the drag continued well into the pattern dot matrix UI area.

The approach was to make the touchable area as wide as the available
space which required a coordinate system transform when calling the
onDrag method of the view-model.

Test: manually verified that drags that start to the left or right, outside of the dot pattern matrix area are picked up and the correct dots are selected
Test: manually verified that password and PIN bouncer layouts are
unaffected
Test: manually verified that all PIN bouncer layouts (folded, unfolded,
half-folded, landscape) are unaffected
Flag: com.android.systemui.scene_container

Change-Id: I9c51a94467cc95c31a5dc7d60f4683901b1e2937