Fix Accessibility Menu Window breaking gesture exclusion API
The gesture exclusion region set by the Accessibility Menu Window counts towards the global exclusion limit. With a few menu items, there is no gesture exclusion margin left to be used by apps.

To fix this, the PRIVATE_FLAG_UNRESTRICTED_GESTURE_EXCLUSION flag is set on the Accessibility Menu Window. This ensures that the exclusion region is not counted towards the global limit.

Additionally, this fixes the problem that the Accessibility Menu Window itself runs out of exclusion margin when it has many menu items. In that case, it's easy to trigger the back gesture when the intention is to drag the window.

Bug: 378425927
Test: Manual, i.e. verified that app's exclusion regions are unaffected by the presence of the Accessibility Menu. Also verified that dragging the Accessibility Menu never triggers the back gesture.
Flag: EXEMPT bugfix
Change-Id: I8f4171b92acfdeffd0e1bc383587320e15c9ef71