Desks: Allow obtaining non-existent desk instance in single-desk setups
An unrelated SystemUI crash might wipe the desk/task state from the
DesktopRepository, and upon recreation, the existing freeform task
recreation might be faster than the desk recreation. This resulted in a
crash-loop because DesktopRepository was enforcing that a desk had been
created before attempting to add a task to it.

That enforcenment isn't really necessary in single-desk devices, so this
CL removes it to fix the regression.

Flag: EXEMPT bug fix
Test: crash SystemUI on purpose, check it recovers instead of getting
into a crash-loop
Bug: 391984373

Change-Id: I37a4a91e06d77207ff3a8da52ae14a7ad30e3616