Avoid a bitmap error in robolectric tests & fix a small UI flaw
The behavior for rendering the dot on the bubble overflow was slightly
different from how it's done on normal bubbles. This CL aligns them.

Additionally, not calling this normalizer get scale function avoids
a robolectric error about bitmaps when using BubbleStackView in tests
which is the true motivation for this change.

Flag: EXEMPT trivial bug fix
Test: atest BubbleControllerTest (with other CL)
Test: manual - get the dot to show on the bubble overflow:
               1) create a bunch of bubbles, don't open them, dismiss
                  the stack
               2) get two bubbles, expand the stack
               => Observe that the overflow is rendered correctly and
                  the "dot" indicating un-opened bubbles is matching
                  with the position of the bubble with a dot
Bug: 375513387
Change-Id: I8d9e2e2de33675883bd3297e6fcb6ad87f95c577