Handle HSUM in OverlayManager shell commands
In HSUM the default user is not USER_SYSTEM, but the OverlayManager
shell commands always defaulted to this user.

This handles HSUM mode by defaulting to USER_CURRENT and resolving it to
the actual user id. This fixes the output of commands such as
"cmd overlay dump --user current state <packageName>" to no longer
include the additional defaultOverlays line, when ran in HSUM mode.

The InstallOverlayTests were also affected by HSUM mode, as the packages
were installed and uninstalled for USER_SYSTEM by default. This also
fixes the test suite by specifing the userId of the current user.

Additionally removes the mention of the userId arg in the fabricate
shell command helper, as this command is always applied for all users.

Flag: EXEMPT bugfix
Bug: 389870829
Bug: 385607637
Bug: 383212992
Test: atest InstallOverlayTests
Change-Id: Ib2ef6e3078ae6970e90b5f58507dfce7a6516a49