Remove Modifier.measureTracer {}
This CL removes Modifier.measureTracer {} given that it was unused and
most importantly misleading. From adamp@:

"The layout at the end returns an object that contains the placement
code as a callback. Many layouts measure their children as part of
placement, not container measurement, because the child's measured size
cannot affect the measurement of the parent. This aids compose in
skipping unnecessary layout up the hierarchy.There are many ways that a
child of this modifier can be remeasured completely outside of this
trace recording."

Bug: 291071158
Test: N/A
Flag: EXEMPT deleted unused code
Change-Id: I9cc6a728e3aa90615dfd1fd7e0a700e52f3319a9