Wrap clock views in FrameLayout when attaching to AndroidView
This allows the underlying view to be changed out w/o needing to force
AndroidView to invoke factory again. Previously I used `key` to force
that recompose to happen whenever the view changed, but this broke the
shared element clock transitions.

Using an intermediate FrameLayout achieves the same result. It's not my
preferred solution as we've had issues in the past that have been
annoying to track down due to layers of parent views, but this solves
the issue, aligns better with the way AndroidView is usually called,
and shouldn't create any new bugs with clipping disabled.

Bug: 393464939
Flag: com.android.systemui.scene_container
Test: Added code to force repro then validated bug was fixed
Change-Id: Ic85e86ced1f496536ddace9dea896cf12a66bff6