[13/N] Desks: Implement move to fullscreen from a desk
Updates the move-to-fullscreen interactions (App Header button, app
handle, etc), to properly move a desk's task out of the desk and into
fullscreen. This includes reparenting the task to the default TDA,
tracking the transition and updating the repository state when the
transition is ready to remove the task from the desk and deactivate the
desk.

Also wires the adb command to move a task out of desktop to this new
implementation.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 393962404
Bug: 393977227
Bug: 394268248
Test: drag task to desktop, then drag back to fullscreen

Change-Id: I9335b8cf796f1ee6062c9c1f50b775bade2c0abf